Scripts/requirements.txt
pandas
python-dotenv
```